Move `ExperimentControl` to `execution`
