+++ packages/engine/lib/execution/src/package/mod.rs
@@ -1,1 +1,2 @@
+pub mod experiment;
 pub mod simulation;
