+++ packages/engine/src/experiment/controller/controller.rs
@@ -1,6 +1,7 @@
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
 use execution::{
+    package::experiment::comms::ExperimentControl,
     runner::comms::{DatastoreSimulationPayload, ExperimentInitRunnerMsgBase, NewSimulationRun},
     worker_pool::comms::{
         experiment::{ExpMsgSend, ExperimentToWorkerPoolMsg},
@@ -27,7 +28,6 @@ use crate::{
             sim_configurer::SimConfigurer,
         },
         package::{ExperimentPackageComms, StepUpdate},
-        ExperimentControl,
     },
     output::OutputPersistenceCreatorRepr,
     proto::{EngineMsg, EngineStatus, ExperimentRunTrait},
