+++ packages/engine/src/types.rs
@@ -2,5 +2,3 @@
 pub const BATCH_ID_LENGTH: usize = 32;
 // TODO: UNUSED: Needs triage
 pub type BatchId = [u8; BATCH_ID_LENGTH];
-
-pub type SpanId = Option<tracing::span::Id>;
